refactor(extension): Use a more explicit naming convention.
